% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R, R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{World Health Organization TB data}
\format{
## `fire`
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{distance from firestation in miles}
  \item{DAMAGE}{x1000 dollars in damages}
  ...
}

## `fire`
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{distance from firestation in miles}
  \item{DAMAGE}{x1000 dollars in damages}
  ...
}
}
\source{
<https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850>

<https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850>
}
\usage{
fire

fire
}
\description{
A subset of data about the damage amount in thousands to Houses that catch fire, along with the how far away the House was to anyfirestation.
this was sampled from 15 houses in a suburban part of a major city

A subset of data about the damage amount in thousands to Houses that catch fire, along with the how far away the House was to anyfirestation.
this was sampled from 15 houses in a suburban part of a major city
}
\keyword{datasets}
