% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbin.R
\name{logbin}
\alias{logbin}
\title{Log-Likelihood Function for the Binomial Distribution}
\usage{
logbin(x, param)
}
\arguments{
\item{x}{An integer or numeric vector of observed values (number of successes).}

\item{param}{A numeric value representing the probability of success (between 0 and 1).}
}
\value{
A numeric vector of log-likelihood values.
}
\description{
Computes the log-likelihood for a binomial random variable with a fixed size of 10
and a given probability parameter.
}
