% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH4753JASON.R
\docType{package}
\name{MATH4753JASON}
\alias{MATH4753JASON-package}
\alias{MATH4753JASON}
\title{TSS package}
\description{
Contains one function from lab3 that creates the total sum of sqaures
}
\author{
Jason Salinas
}
