% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson.R
\name{maxlikg2}
\alias{maxlikg2}
\title{Poisson Likelihood}
\usage{
maxlikg2(theta1, theta2, lfun = "logbinpois", ...)
}
\arguments{
\item{theta1}{Gets the first theta given}

\item{theta2}{Gets the second theta given}

\item{lfun}{Gets the function that will be used (going to have to create your own)}

\item{...}{Anything else can be added so things like title and moving the text etc.}
}
\value{
A plot of what the Poisson likelihood will look like given the theta's and function
}
\description{
Poisson Likelihood
}
