% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Confidence Interval Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{num  of bootstrap iterations}

\item{x}{data sample to bootstrap}

\item{fun}{function given}

\item{alpha}{Significance level}

\item{cx}{character expansion factor to make text look better}

\item{...}{more elements passed into other functions in use like histogram}
}
\value{
An  list containing:
\describe{
  \item{ci}{the confidence interval}
  \item{fun}{ function used for bootstrap}
  \item{x}{input sample}
}
}
\description{
This function performs  bootstrapping on a given numeric sample to estimate
the sampling distribution of a statisti and calculates a confidence interval
}
