% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{My clytb}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{Sample size for each iteration.}

\item{iter}{Number of iterations}

\item{p}{probability of success in the binomial test
.}
}
\value{
A histogram with an overlaid normal curve.
}
\description{
My clytb
}
