% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyCLTP.R
\name{mycltp}
\alias{mycltp}
\title{The Poisson Sample}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{gets the sample size}

\item{iter}{gets the amount of times the function will iterate}

\item{lambda}{represents the Poisson distribution which is defaulted to 10}

\item{...}{any values can be added onto here so maybe things like graphical design.}
}
\value{
a graph showing the n, iteration, and poisson distribution
}
\description{
The Poisson Sample
}
