% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{Maximum Likelihood Plotter for 1D Likelihood Functions}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{A log-likelihood function that takes a single numeric parameter as input.
The default is `"logbin2"` which must be defined in the environment.}

\item{theta}{A numeric vector of parameter values to evaluate the likelihood over.}
}
\value{
The value of `theta` that maximizes the likelihood.
}
\description{
Evaluates a log-likelihood function over a range of parameter values and plots the likelihood curve (exponentiated log-likelihood).
The function highlights the parameter value that maximizes the likelihood.
}
\details{
The function:
\itemize{
  \item Evaluates the log-likelihood function at each value of `theta`.
  \item Converts the log-likelihood to standard likelihood via exponentiation.
  \item Plots the likelihood curve and adds a vertical line at the maximum.
  \item Displays the value of `theta` that gives the highest likelihood.
}
}
\examples{
# Example with a user-defined binomial log-likelihood function
logbin2 <- function(p) dbinom(5, size = 10, prob = p, log = TRUE)
mymaxlikg(lfun = logbin2, theta = seq(0.01, 0.99, length.out = 100))

}
