% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{computes the area of curve}
\usage{
myncurve(mu, sigma, doh)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard devetion}

\item{doh}{value up to which the area under the curve}
}
\value{
list of mean and sd and can be made into a curve
}
\description{
computes the area of curve
}
