% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{A quadratic function that plots}
\usage{
myplot(x)
}
\arguments{
\item{x}{is a numeric value}
}
\value{
a numeric value that is quadratic
}
\description{
A quadratic function that plots
}
