% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePlot.R
\name{mysample}
\alias{mysample}
\title{timePlot}
\usage{
mysample(n = 1000, iter = 10, time = 0.5)
}
\arguments{
\item{n}{a numeric value (default value of 1000)}

\item{iter}{a numeric value that iterates (default value of 10)}

\item{time}{a numeric value that determines the time (default value of 0.5)}
}
\value{
a animation of the samples being plotted based on the iter, n, and time value.
}
\description{
timePlot
}
