% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project1.R
\name{ntickets}
\alias{ntickets}
\title{Discrete and Continuous ticket probability}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{gets the number of seats in flight}

\item{gamma}{gets the probability that the plane will be overbooked}

\item{p}{gets the probability that the person will actually show up}
}
\value{
two charts with a continuous and discrete plotting of where the optimal tickets should be sold
}
\description{
Discrete and Continuous ticket probability
}
