% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.R
\name{computeTS}
\alias{computeTS}
\title{computes the Total sum of squares using a data set and a responce var}
\usage{
computeTS(data, response_var)
}
\arguments{
\item{data}{the data set}

\item{response_var}{A Var in a data set}
}
\value{
A numeric vector which represents the Total sum of squares
}
\description{
computes the Total sum of squares using a data set and a responce var
}
