% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIREDAM.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM data set for fire damage and distance from fire station.}
\format{
A data set with 15 rows and 2 variables:
\describe{
  \item{FIRE DAMAGE}{Cost of fire damage (in thousands of dollars)}
  \item{DISTANCe}{Distance of the fire station to the fire (in miles)}
  ...
}
}
\source{
Mendenhall, William M., and Terry Sincich. Statistics for Engineering and the Sciences. CRC Press, Taylor & Francis Group, 2016, pg. 541.
}
\usage{
fire
}
\description{
15 lines of measurements of the cost of fire damage (in thousands of dollars) and the distance of the fire to the fire station (in miles)
}
\keyword{datasets}
