% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{this value is basically the amount of times the sample will loop}

\item{x}{a vector of the data set}

\item{fun}{the function used for the bootstrap (example mean, median, etc)}

\item{alpha}{the alpha value is used to help determine quantile}

\item{cx}{the cx value is basically the controlled x value}

\item{...}{this can accept any values with things such as xlab}
}
\value{
this will return a bootstrap histogram of the values given through the function
}
\description{
Bootstrap
}
