% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{myboot2}
\alias{myboot2}
\title{BootStrap}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterates a number of times given (or 10000 by default)}

\item{x}{gets the sample (random)}

\item{fun}{determines which function to use with mean being default}

\item{alpha}{determines the confidence interval with 0.05 being default (95% confidence)}

\item{cx}{determines position in terms of x (1.5 by default)}

\item{...}{anything else can be added to this such as xlab and more}
}
\value{
a histogram of the bootstrap samples given
}
\description{
BootStrap
}
