% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypergeo.R
\name{myhyper}
\alias{myhyper}
\title{function simulates the Hypergeometric distribution by,
Drawing iter samples of size n without replacement from a population of size N that contains r successes.}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{The number of iterations}

\item{N}{The population size}

\item{r}{The number of success states in the population}

\item{n}{The sample size}
}
\value{
barplot of the proportions hyper function
}
\description{
function simulates the Hypergeometric distribution by,
Drawing iter samples of size n without replacement from a population of size N that contains r successes.
}
