% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCurve.R
\name{myncurve}
\alias{myncurve}
\title{Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{gets given mean}

\item{sigma}{gets given sigma}

\item{a}{gets a (which is x)}
}
\value{
a plot of the given values
}
\description{
Curve
}
